Make the Plugin handle MESSAGE and EVENT hooks (we want both).  Add start of Replies and Routines
Replies are the RPL_ stuff in RFC1459

Routines are a sequence of messages, events, etc.  Supposed to be an abstract concept for, well, a "routine" of the server.

I am still having issues trying to abstract this out properly.  I want to over-engineer this in a way that makes sense and is workable, and I'm still struggling with that at this point in the process.

I think more implementation of features, as opposed to high level thinking, will result in a better working model.  Will continue to refactor as I begin to understand the needs of the plugin (such as the Event/Message split in the event system)